Add support for legacy (0.x) serialization
* If config "use-legacy-serialization" set:
  * Store `PersistentRepr` whole in payload as bin
  * Fully delegate serialization to akka-ser
  * Use type = 'repr'

Also
* Fix numeric widening warning
* Remove old .gitkeep file
* Go to 5x in akka dilation for travis

+++ casbah/src/main/scala/akka/contrib/persistence/mongodb/CasbahPersistenceReadJournaller.scala
@@ -46,7 +46,7 @@ class AllEvents(val driver: CasbahMongoDriver) extends SyncActorPublisher[EventE
           .flatMap(lst => lst.collect {case x:DBObject => x} )
           .map(driver.deserializeJournal)
           .zipWithIndex
-          .map { case(e,i) => e.toEnvelope(i) }
+          .map { case(e,i) => e.toEnvelope(i.toLong) }
 
   override protected def next(c: Stream[EventEnvelope], atMost: Long): (Vector[EventEnvelope], Stream[EventEnvelope]) = {
     val (buf,remainder) = c.splitAt(atMost.toIntWithoutWrapping)
@@ -76,7 +76,7 @@ class EventsByPersistenceId(val driver: CasbahMongoDriver, persistenceId: String
       .filter(dbo => dbo.getAs[Long](SEQUENCE_NUMBER).exists(sn => sn >= fromSeq && sn <= toSeq))
       .map(driver.deserializeJournal)
       .zipWithIndex
-      .map { case(e,i) => e.toEnvelope(i) }
+      .map { case(e,i) => e.toEnvelope(i.toLong) }
 
   override protected def next(c: Stream[EventEnvelope], atMost: Long): (Vector[EventEnvelope], Stream[EventEnvelope]) = {
     val (buf,remainder) = c.splitAt(atMost.toIntWithoutWrapping)
